Implementation of Restore. Refactoring P2DataAccess and P2Facet
