+++ nexus-repository-p2/src/main/java/org/sonatype/nexus/repository/p2/internal/P2FacetImpl.java
@@ -12,17 +12,35 @@
  */
 package org.sonatype.nexus.repository.p2.internal;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.List;
 import java.util.Map;
 
+import javax.annotation.Nullable;
 import javax.inject.Named;
 
+import com.google.common.base.Supplier;
+import com.google.common.collect.ImmutableList;
+import org.sonatype.nexus.blobstore.api.Blob;
+import org.sonatype.nexus.common.collect.AttributesMap;
+import org.sonatype.nexus.common.hash.HashAlgorithm;
 import org.sonatype.nexus.repository.FacetSupport;
+import org.sonatype.nexus.repository.Repository;
 import org.sonatype.nexus.repository.p2.P2Facet;
 import org.sonatype.nexus.repository.storage.Asset;
+import org.sonatype.nexus.repository.storage.AssetBlob;
 import org.sonatype.nexus.repository.storage.Bucket;
 import org.sonatype.nexus.repository.storage.Component;
+import org.sonatype.nexus.repository.storage.MetadataNodeEntityAdapter;
+import org.sonatype.nexus.repository.storage.Query;
 import org.sonatype.nexus.repository.storage.StorageTx;
+import org.sonatype.nexus.repository.view.Content;
+import org.sonatype.nexus.repository.view.Payload;
+import org.sonatype.nexus.repository.view.payloads.BlobPayload;
 
+import static java.util.Collections.singletonList;
+import static org.sonatype.nexus.common.hash.HashAlgorithm.SHA1;
 import static org.sonatype.nexus.repository.p2.internal.AssetKind.ARTIFACT_JAR;
 import static org.sonatype.nexus.repository.p2.internal.AssetKind.ARTIFACT_XML;
 import static org.sonatype.nexus.repository.p2.internal.AssetKind.ARTIFACT_XML_XZ;
@@ -44,8 +62,6 @@ import static org.sonatype.nexus.repository.p2.internal.proxy.P2ProxyRecipe.CONT
 import static org.sonatype.nexus.repository.p2.internal.proxy.P2ProxyRecipe.JAR_EXTENSION;
 import static org.sonatype.nexus.repository.p2.internal.proxy.P2ProxyRecipe.XML_EXTENSION;
 import static org.sonatype.nexus.repository.p2.internal.proxy.P2ProxyRecipe.XML_XZ_EXTENSION;
-import static org.sonatype.nexus.repository.p2.internal.util.P2DataAccess.findAsset;
-import static org.sonatype.nexus.repository.p2.internal.util.P2DataAccess.findComponent;
 import static org.sonatype.nexus.repository.p2.internal.util.P2PathUtils.PLUGIN_NAME;
 import static org.sonatype.nexus.repository.storage.AssetEntityAdapter.P_ASSET_KIND;
 import static org.sonatype.nexus.repository.storage.ComponentEntityAdapter.P_VERSION;
@@ -61,6 +77,8 @@ public class P2FacetImpl
     extends FacetSupport
     implements P2Facet
 {
+  public static final List<HashAlgorithm> HASH_ALGORITHMS = ImmutableList.of(SHA1);
+
   @Override
   public Component findOrCreateComponent(final StorageTx tx, final String path, final Map<String, String> attributes) {
     String name = attributes.get(P_NAME);
@@ -121,6 +139,90 @@ public class P2FacetImpl
     }
 
     return asset;
+  }
+
+  /**
+   * Find a component by its name and tag (version)
+   *
+   * @return found component of null if not found
+   */
+  @Nullable
+  public Component findComponent(final StorageTx tx,
+                                        final Repository repository,
+                                        final String name,
+                                        final String version)
+  {
+    Iterable<Component> components = tx.findComponents(
+        Query.builder()
+            .where(P_NAME).eq(name)
+            .and(P_VERSION).eq(version)
+            .build(),
+        singletonList(repository)
+    );
+    if (components.iterator().hasNext()) {
+      return components.iterator().next();
+    }
+    return null;
+  }
+
+  /**
+   * Find an asset by its name.
+   *
+   * @return found asset or null if not found
+   */
+  @Nullable
+  public Asset findAsset(final StorageTx tx, final Bucket bucket, final String assetName) {
+    return tx.findAssetWithProperty(MetadataNodeEntityAdapter.P_NAME, assetName, bucket);
+  }
+
+  /**
+   * Save an asset && create blob.
+   *
+   * @return blob content
+   */
+  public Content saveAsset(final StorageTx tx,
+                                  final Asset asset,
+                                  final Supplier<InputStream> contentSupplier,
+                                  final Payload payload) throws IOException
+  {
+    AttributesMap contentAttributes = null;
+    String contentType = null;
+    if (payload instanceof Content) {
+      contentAttributes = ((Content) payload).getAttributes();
+      contentType = payload.getContentType();
+    }
+    return saveAsset(tx, asset, contentSupplier, contentType, contentAttributes);
+  }
+
+  /**
+   * Save an asset && create blob.
+   *
+   * @return blob content
+   */
+  public Content saveAsset(final StorageTx tx,
+                                  final Asset asset,
+                                  final Supplier<InputStream> contentSupplier,
+                                  final String contentType,
+                                  @Nullable final AttributesMap contentAttributes) throws IOException
+  {
+    Content.applyToAsset(asset, Content.maintainLastModified(asset, contentAttributes));
+    AssetBlob assetBlob = tx.setBlob(
+        asset, asset.name(), contentSupplier, HASH_ALGORITHMS, null, contentType, false
+    );
+    asset.markAsDownloaded();
+    tx.saveAsset(asset);
+    return toContent(asset, assetBlob.getBlob());
+  }
+
+  /**
+   * Convert an asset blob to {@link Content}.
+   *
+   * @return content of asset blob
+   */
+  public Content toContent(final Asset asset, final Blob blob) {
+    Content content = new Content(new BlobPayload(blob, asset.requireContentType()));
+    Content.extractFromAsset(asset, HASH_ALGORITHMS, content.getAttributes());
+    return content;
   }
 
   @Override
