+++ nexus-repository-p2/src/main/java/org/sonatype/nexus/repository/p2/internal/util/P2DataAccess.java
@@ -13,43 +13,19 @@
 package org.sonatype.nexus.repository.p2.internal.util;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.List;
 import java.util.Optional;
 
-import javax.annotation.Nullable;
 import javax.inject.Inject;
 import javax.inject.Named;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.slf4j.Logger;
 import org.sonatype.goodies.common.Loggers;
-import org.sonatype.nexus.blobstore.api.Blob;
-import org.sonatype.nexus.common.collect.AttributesMap;
-import org.sonatype.nexus.common.hash.HashAlgorithm;
-import org.sonatype.nexus.repository.Repository;
 import org.sonatype.nexus.repository.p2.internal.exception.AttributeParsingException;
 import org.sonatype.nexus.repository.p2.internal.metadata.P2Attributes;
-import org.sonatype.nexus.repository.storage.Asset;
-import org.sonatype.nexus.repository.storage.AssetBlob;
-import org.sonatype.nexus.repository.storage.Bucket;
-import org.sonatype.nexus.repository.storage.Component;
-import org.sonatype.nexus.repository.storage.MetadataNodeEntityAdapter;
-import org.sonatype.nexus.repository.storage.Query;
-import org.sonatype.nexus.repository.storage.StorageTx;
 import org.sonatype.nexus.repository.storage.TempBlob;
-import org.sonatype.nexus.repository.view.Content;
-import org.sonatype.nexus.repository.view.Payload;
-import org.sonatype.nexus.repository.view.payloads.BlobPayload;
-
-import com.google.common.base.Supplier;
-import com.google.common.collect.ImmutableList;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static java.util.Collections.singletonList;
-import static org.sonatype.nexus.common.hash.HashAlgorithm.SHA1;
-import static org.sonatype.nexus.repository.storage.ComponentEntityAdapter.P_VERSION;
-import static org.sonatype.nexus.repository.storage.MetadataNodeEntityAdapter.P_NAME;
 
 /**
  * Shared code between P2 facets.
@@ -58,99 +34,17 @@ import static org.sonatype.nexus.repository.storage.MetadataNodeEntityAdapter.P_
 public class P2DataAccess
 {
   final Logger log = Loggers.getLogger(this);
-  public static final List<HashAlgorithm> HASH_ALGORITHMS = ImmutableList.of(SHA1);
 
   private final AttributesParserFeatureXml featureXmlParser;
+
   private final AttributesParserManifest manifestParser;
 
   @Inject
-  public P2DataAccess(final AttributesParserFeatureXml featureXmlParser, final AttributesParserManifest manifestParser) {
+  public P2DataAccess(final AttributesParserFeatureXml featureXmlParser,
+                      final AttributesParserManifest manifestParser)
+  {
     this.featureXmlParser = checkNotNull(featureXmlParser);
     this.manifestParser = checkNotNull(manifestParser);
-  }
-
-  /**
-   * Find a component by its name and tag (version)
-   *
-   * @return found component of null if not found
-   */
-  @Nullable
-  public static Component findComponent(final StorageTx tx,
-                                 final Repository repository,
-                                 final String name,
-                                 final String version)
-  {
-    Iterable<Component> components = tx.findComponents(
-        Query.builder()
-            .where(P_NAME).eq(name)
-            .and(P_VERSION).eq(version)
-            .build(),
-        singletonList(repository)
-    );
-    if (components.iterator().hasNext()) {
-      return components.iterator().next();
-    }
-    return null;
-  }
-
-  /**
-   * Find an asset by its name.
-   *
-   * @return found asset or null if not found
-   */
-  @Nullable
-  public static Asset findAsset(final StorageTx tx, final Bucket bucket, final String assetName) {
-    return tx.findAssetWithProperty(MetadataNodeEntityAdapter.P_NAME, assetName, bucket);
-  }
-
-  /**
-   * Save an asset && create blob.
-   *
-   * @return blob content
-   */
-  public static Content saveAsset(final StorageTx tx,
-                           final Asset asset,
-                           final Supplier<InputStream> contentSupplier,
-                           final Payload payload) throws IOException
-  {
-    AttributesMap contentAttributes = null;
-    String contentType = null;
-    if (payload instanceof Content) {
-      contentAttributes = ((Content) payload).getAttributes();
-      contentType = payload.getContentType();
-    }
-    return saveAsset(tx, asset, contentSupplier, contentType, contentAttributes);
-  }
-
-  /**
-   * Save an asset && create blob.
-   *
-   * @return blob content
-   */
-  public static Content saveAsset(final StorageTx tx,
-                           final Asset asset,
-                           final Supplier<InputStream> contentSupplier,
-                           final String contentType,
-                           @Nullable final AttributesMap contentAttributes) throws IOException
-  {
-    Content.applyToAsset(asset, Content.maintainLastModified(asset, contentAttributes));
-    AssetBlob assetBlob = tx.setBlob(
-        asset, asset.name(), contentSupplier, HASH_ALGORITHMS, null, contentType, false
-    );
-    asset.markAsDownloaded();
-    tx.saveAsset(asset);
-    return toContent(asset, assetBlob.getBlob());
-  }
-
-  /**
-   * Convert an asset blob to {@link Content}.
-   *
-   * @return content of asset blob
-   */
-  public static Content toContent(final Asset asset, final Blob blob) {
-    Content content = new Content(new BlobPayload(blob, asset.requireContentType()));
-    Content.extractFromAsset(asset, HASH_ALGORITHMS, content.getAttributes());
-    return content;
   }
 
   @VisibleForTesting
@@ -173,7 +67,7 @@ public class P2DataAccess
     }
 
     return Optional.ofNullable(p2Attributes)
-        .filter(jarP2Attributes-> !jarP2Attributes.isEmpty())
+        .filter(jarP2Attributes -> !jarP2Attributes.isEmpty())
         .map(jarP2Attributes -> P2Attributes.builder().merge(sourceP2Attributes, jarP2Attributes).build())
         .orElse(sourceP2Attributes);
   }
